refactor: move GemmaTermsDialog to shared/popups directory and clean up debug logs
